migrate player with notes to lesson page
