+++ src/components/pages/lessons/transcript.tsx
@@ -1,16 +1,12 @@
-import React, {FunctionComponent, useState, useEffect} from 'react'
-import {animateScroll as scroll} from 'react-scroll'
-import {get, first, noop} from 'lodash'
+import React, {FunctionComponent} from 'react'
+
 import ReactMarkdown from 'react-markdown'
 import CodeBlock from 'components/code-block'
 
 type TranscriptProps = {
-  player: any
   className?: string
-  playVideo: () => any
   initialTranscript?: string
   enhancedTranscript: string
-  playerAvailable: boolean
 }
 
 const hmsToSeconds = (str: string) => {
@@ -26,68 +22,15 @@ const hmsToSeconds = (str: string) => {
 }
 
 // https://regexr.com/58bnr
-const regex = /[0-9]:[0-9][0-9]|[0-9]{2}:[0-9][0-9]|[[0-9]{2}:[0-9][0-9]]|[[0-9]{3}:[0-9][0-9]]/g
+const regex =
+  /[0-9]:[0-9][0-9]|[0-9]{2}:[0-9][0-9]|[[0-9]{2}:[0-9][0-9]]|[[0-9]{3}:[0-9][0-9]]/g
 
 const Transcript: FunctionComponent<TranscriptProps> = ({
-  player,
   className,
-  playVideo = noop,
   initialTranscript = '',
   enhancedTranscript,
-  playerAvailable,
 }: TranscriptProps) => {
-  const transcriptText = enhancedTranscript || initialTranscript
-  const [transcript, setTranscript] = useState<string>()
-
-  const LinkReference = (props: any) => {
-    const children = get(props, 'children', [''])
-    const linkText: any = first(children)
-    const value: string = linkText.props.children
-    const number: string = value.replace('[', '').replace(']', '')
-    const secondsToSeek = hmsToSeconds(number)
-
-    if (value.match(regex)) {
-      return (
-        <button
-          className="text-blue-600 hover:underline"
-          onClick={() => {
-            const duration = player.current.getDuration()
-            const fractionToSeek = secondsToSeek / duration
-            player.current.seekTo(fractionToSeek)
-            playVideo()
-            scroll.scrollToTop({
-              duration: 300,
-              smooth: 'true',
-            })
-          }}
-        >
-          {number}
-        </button>
-      )
-    }
-    return <a href={props.href}>{value}</a>
-  }
-
-  useEffect(() => {
-    if (transcriptText && playerAvailable) {
-      const matches = transcriptText && transcriptText.match(regex)
-      let result = transcriptText
-      matches &&
-        matches.forEach((match: any, i: number) => {
-          result = result.replace(
-            match,
-            `[[${match.replace('[', '').replace(']', '')}]](${match
-              .replace('[', '')
-              .replace(']', '')})`,
-          )
-          if (i === matches.length - 1) {
-            setTranscript(result)
-          }
-        })
-    } else {
-      setTranscript(transcriptText)
-    }
-  }, [transcriptText, playerAvailable, initialTranscript])
+  const transcript = enhancedTranscript || initialTranscript
 
   if (!transcript) {
     return null
@@ -98,7 +41,6 @@ const Transcript: FunctionComponent<TranscriptProps> = ({
       <ReactMarkdown
         skipHtml={false}
         renderers={{
-          link: LinkReference,
           code: (props) => {
             return <CodeBlock {...props} />
           },
