+++ src/components/player/index.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react'
-import classNames from 'classnames'
 import {isEmpty} from 'lodash'
 import {
   Player,
@@ -22,23 +21,148 @@ import {
 import HLSSource from './hls-source'
 import CueBar from './cue-bar'
 import ControlBarDivider from './control-bar-divider'
-import {Tab, TabList, TabPanel, TabPanels, Tabs} from '@reach/tabs'
-import CollectionLessonsList from 'components/pages/lessons/collection-lessons-list'
-import {useEggheadPlayerPrefs} from 'components/EggheadPlayer/use-egghead-player'
-import SimpleBar from 'simplebar-react'
-import {Element} from 'react-scroll'
-import ReactMarkdown from 'react-markdown'
-import {convertTime} from 'utils/time-utils'
+import {
+  defaultSubtitlePreference,
+  useEggheadPlayerPrefs,
+} from 'components/EggheadPlayer/use-egghead-player'
 import {VideoResource} from 'types'
+import {MutableRefObject, SyntheticEvent} from 'react'
+import noop from 'utils/noop'
+import useBreakpoint from 'utils/breakpoints'
 
-const VideoResourcePlayer: React.FC<{videoResource: VideoResource}> = ({
+export type VideoResourcePlayerProps = {
+  videoResource: VideoResource
+  containerRef?: MutableRefObject<any>
+  onCanPlay?: (event?: SyntheticEvent) => void
+  onPause?: () => void
+  onPlay?: () => void
+  onTimeUpdate?: (event: any) => void
+  onEnded?: () => void
+  onVolumeChange?: (event?: SyntheticEvent) => void
+  hidden?: boolean
+  className?: string
+  volume?: number
+}
+
+const VideoResourcePlayer: React.FC<VideoResourcePlayerProps> = ({
   videoResource,
+  containerRef,
+  hidden = false,
+  className,
+  children,
+  ...props
 }) => {
-  const playerContainer = React.useRef<any>()
   const {setPlayerPrefs, getPlayerPrefs} = useEggheadPlayerPrefs()
-  const {activeSidebarTab} = getPlayerPrefs()
+  const {hasNotes} = useNotesCues(videoResource)
+  const {subtitle, playbackRate} = getPlayerPrefs()
   const {player} = usePlayer()
+  const {sm} = useBreakpoint()
 
+  return (
+    <div
+      className={`relative z-10 ${className} 
+          ${hidden ? 'hidden' : 'block'} 
+          ${!hasNotes ? 'pb-14' : 'pb-[4.5rem]'} 
+          ${player.isFullscreen ? 'lg:col-span-12' : 'lg:col-span-9'}`}
+    >
+      <Player
+        muted
+        autoplay
+        crossOrigin="anonymous"
+        className="font-sans"
+        volume={0.2}
+        poster={videoResource.thumb_url}
+        {...props}
+      >
+        <BigPlayButton position="center" />
+        {videoResource.hls_url && (
+          <HLSSource isVideoChild src={videoResource.hls_url} />
+        )}
+        {videoResource.subtitles_url && (
+          <track
+            key={videoResource.subtitles_url}
+            src={videoResource.subtitles_url}
+            kind="subtitles"
+            srcLang="en"
+            label="English"
+            default={subtitle?.language === 'en'}
+          />
+        )}
+        {hasNotes && (
+          <track
+            id="notes"
+            src={`/api/github-load-notes?url=${videoResource.staff_notes_url}`}
+            kind="metadata"
+            label="notes"
+          />
+        )}
+        {!sm && <CueBar key="cue-bar" order={6.0} />}
+        <ProgressControl key="progress-control" order={7.0} />
+        {!sm && (
+          <ControlBar
+            disableDefaultControls
+            autoHide={false}
+            className={`transform ${
+              !hasNotes ? 'translate-y-14' : 'translate-y-[4.5rem]'
+            }`}
+            order={8.0}
+          >
+            <PlayToggle key="play-toggle" order={1} />
+            <ReplayControl key="replay-control" order={2} />
+            <ForwardControl key="forward-control" order={3} />
+            <VolumeMenuButton key="volume-menu-button" order={4} />
+            <CurrentTimeDisplay key="current-time-display" order={5} />
+            <TimeDivider key="time-divider" order={6} />
+            <DurationDisplay key="duration-display" order={7} />
+            <ControlBarDivider key="divider" order={9} className="flex-grow" />
+            <RemainingTimeDisplay key="remaining-time-display" order={10} />
+            <PlaybackRateMenuButton
+              rates={[1, 1.25, 1.5, 2]}
+              key="playback-rate"
+              order={11}
+              selected={playbackRate}
+              onChange={(playbackRate: number) => {
+                setPlayerPrefs({playbackRate})
+              }}
+            />
+            {videoResource.subtitles_url && (
+              <ClosedCaptionButton
+                key={videoResource.subtitles_url}
+                order={12}
+                selected={subtitle}
+                onChange={(track?: TextTrack) => {
+                  const updatedSubtitlePref = track
+                    ? {
+                        id: track.id,
+                        kind: track.kind,
+                        label: track.label,
+                        language: track.language,
+                      }
+                    : defaultSubtitlePreference
+
+                  setPlayerPrefs({
+                    subtitle: updatedSubtitlePref,
+                  })
+                }}
+              >
+                1123
+              </ClosedCaptionButton>
+            )}
+            <FullscreenToggle
+              key="fullscreen-toggle"
+              fullscreenElement={containerRef?.current}
+              order={13}
+            />
+          </ControlBar>
+        )}
+      </Player>
+      {children}
+    </div>
+  )
+}
+
+export const useNotesCues = (videoResource: VideoResource) => {
+  const {player} = usePlayer()
   const {activeMetadataTracks = []} = player
 
   const noteTracks = activeMetadataTracks.filter((track: TextTrack) => {
@@ -58,201 +182,10 @@ const VideoResourcePlayer: React.FC<{videoResource: VideoResource}> = ({
 
   const hasNotes = !isEmpty(videoResource?.staff_notes_url)
 
-  return (
-    <PlayerContainer
-      ref={playerContainer}
-      player={player}
-      className="relative grid grid-cols-1 lg:grid-cols-12 font-sans text-base"
-    >
-      <div
-        className={`relative z-10 ${!hasNotes ? 'pb-14' : 'pb-[4.5rem]'} ${
-          player.isFullscreen ? 'lg:col-span-12' : 'lg:col-span-9'
-        }`}
-      >
-        <Player
-          muted
-          autoplay
-          crossOrigin="anonymous"
-          className="font-sans"
-          poster={videoResource.thumb_url}
-        >
-          <BigPlayButton position="center" />
-          <HLSSource isVideoChild src={videoResource.hls_url} />
-          <track
-            src={videoResource.subtitles_url}
-            kind="subtitles"
-            srcLang="en"
-            label="English"
-            default
-          />
-          {hasNotes && (
-            <track
-              id="notes"
-              src={`/api/github-load-notes?url=${videoResource.staff_notes_url}`}
-              kind="metadata"
-              label="notes"
-            />
-          )}
-
-          <CueBar key="cue-bar" order={6.0} />
-          <ControlBar
-            disableDefaultControls
-            autoHide={false}
-            className={`transform ${
-              !hasNotes ? 'translate-y-14' : 'translate-y-[4.5rem]'
-            }`}
-          >
-            <PlayToggle key="play-toggle" order={1} />
-            <ReplayControl key="replay-control" order={2} />
-            <ForwardControl key="forward-control" order={3} />
-            <VolumeMenuButton key="volume-menu-button" order={4} />
-            <CurrentTimeDisplay key="current-time-display" order={5} />
-            <TimeDivider key="time-divider" order={6} />
-            <DurationDisplay key="duration-display" order={7} />
-            <ControlBarDivider key="divider" order={9} className="flex-grow" />
-            <RemainingTimeDisplay key="remaining-time-display" order={10} />
-            <PlaybackRateMenuButton
-              rates={[1, 1.25, 1.5, 2]}
-              key="playback-rate"
-              order={11}
-            />
-            <ClosedCaptionButton order={12}>1123</ClosedCaptionButton>
-            <FullscreenToggle
-              key="fullscreen-toggle"
-              fullscreenElement={playerContainer.current}
-              order={13}
-            />
-          </ControlBar>
-          <ProgressControl key="progress-control" order={8} />
-        </Player>
-      </div>
-      <div className="lg:col-span-3 side-bar">
-        <div className="relative h-full">
-          {/* TODO: remove weird logic that assumes 2 tabs */}
-          <Tabs
-            index={(hasNotes && activeSidebarTab) || 0}
-            onChange={(tabIndex) =>
-              setPlayerPrefs({activeSidebarTab: tabIndex})
-            }
-            className="max-h-[500px] shadow-sm lg:max-h-[none] lg:absolute left-0 top-0 w-full h-full flex flex-col bg-gray-100 dark:bg-gray-1000 text-gray-900 dark:text-white"
-          >
-            <TabList className="relative z-[1] flex-shrink-0">
-              {!isEmpty(videoResource.collection) && <Tab>Lessons</Tab>}
-              {hasNotes && <Tab>Notes</Tab>}
-            </TabList>
-            <TabPanels className="flex-grow relative">
-              <div className="lg:absolute" css={{inset: 0}}>
-                {!isEmpty(videoResource.collection) && (
-                  <TabPanel className="bg-gray-100 dark:bg-gray-1000 w-full h-full">
-                    <CollectionLessonsList
-                      course={videoResource.collection}
-                      currentLessonSlug={videoResource.slug}
-                      progress={[]}
-                    />
-                  </TabPanel>
-                )}
-                {!isEmpty(cues) && (
-                  <TabPanel className="bg-gray-100 dark:bg-gray-1000 w-full h-full">
-                    <NotesTabContent cues={cues} />
-                  </TabPanel>
-                )}
-              </div>
-            </TabPanels>
-          </Tabs>
-        </div>
-      </div>
-    </PlayerContainer>
-  )
-}
-
-const PlayerContainer: React.ForwardRefExoticComponent<any> = React.forwardRef<
-  HTMLDivElement,
-  any
->((props, ref) => {
-  const {player, className, children, ...rest} = props
-  const {paused, hasStarted, waiting, seeking, isFullscreen, userActivity} =
-    player
-
-  return (
-    <div
-      {...rest}
-      ref={ref}
-      className={classNames(
-        {
-          'cueplayer-react-has-started': hasStarted,
-          'cueplayer-react-paused': paused,
-          'cueplayer-react-playing': !paused,
-          'cueplayer-react-waiting': waiting,
-          'cueplayer-react-seeking': seeking,
-          'cueplayer-react-fullscreen': isFullscreen,
-          'cueplayer-react-user-inactive': !userActivity,
-          'cueplayer-react-user-active': userActivity,
-        },
-        'cueplayer-react',
-        className,
-      )}
-    >
-      {children}
-    </div>
-  )
-})
-
-const NotesTabContent: React.FC<{cues: VTTCue[]}> = ({cues}) => {
-  const {player, manager} = usePlayer()
-  const actions = manager.getActions()
-  const disabled: boolean = isEmpty(cues)
-  const scrollableNodeRef: any = React.createRef()
-
-  return disabled ? null : (
-    <SimpleBar
-      forceVisible="y"
-      autoHide={false}
-      scrollableNodeProps={{
-        ref: scrollableNodeRef,
-        id: 'notes-tab-scroll-container',
-      }}
-      className="h-full overscroll-contain p-4"
-    >
-      <div className="space-y-3">
-        {cues.map((cue: VTTCue) => {
-          const note = cue.text
-          const active = player.activeMetadataTrackCues.includes(cue)
-          return (
-            <div key={cue.startTime}>
-              {active && <Element name="active-note" />}
-              <div
-                className={classNames(
-                  'text-sm p-4 bg-white dark:bg-gray-900 rounded-md shadow-sm border-2 border-transparent',
-                  {
-                    'border-indigo-500': active,
-                    '': !active,
-                  },
-                )}
-              >
-                {note && (
-                  <ReactMarkdown className="leading-normal prose-sm prose dark:prose-dark">
-                    {note}
-                  </ReactMarkdown>
-                )}
-                {cue.startTime && (
-                  <div
-                    onClick={() => {
-                      actions.seek(cue.startTime)
-                    }}
-                    className="w-full cursor-pointer underline flex items-baseline justify-end pt-3 text-gray-900 dark:text-white"
-                  >
-                    <time className="text-xs opacity-60 font-medium">
-                      {convertTime(cue.startTime)}
-                    </time>
-                  </div>
-                )}
-              </div>
-            </div>
-          )
-        })}
-      </div>
-    </SimpleBar>
-  )
+  return {
+    hasNotes,
+    cues,
+  }
 }
 
 export default VideoResourcePlayer
